Breaking out model configuration into its own file. Restructuring testing to address that change.
