+++ test/fixtures.py
@@ -5,8 +5,8 @@ import pytest
 
 @pytest.fixture()
 def setUp():
-    test_dir = os.getcwd()
-    root_dir = os.path.dirname(test_dir)
+    root_dir = os.getcwd()
+    test_dir = os.path.join(root_dir, "test")
     model_dir = os.path.join(root_dir, "models")
     return test_dir, root_dir, model_dir
